Related work items: #72

+++ InformacionCiudades.API/Models/users/Student/StudentUpdateProfileDto.cs
@@ -13,16 +13,14 @@ namespace ApiBolsaTrabajoUTN.API.Models.users.Student
         public string? Province { get; set; }
         public string? City { get; set; }
         public string? PhoneNumb { get; set; }
-        //cv
         public int CareerId { get; set; }
         public int ApprovedSubjets { get; set; }
         public string? PlanDeEstudio { get; set; }
         public int CurrentCareerYear { get; set; }
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public Turn Turn { get; set; }
-        public int Average { get; set; }
-        public int AverageWithFails { get; set; }
-        //public List<string> Skils { get; set; }   //*** sin implementar aun
+        public float Average { get; set; }
+        public float AverageWithFails { get; set; }
         public bool FirstChargeData { get; set; }
         public bool ActiveAccount { get; set; } = true;
     }
